Add files via upload
This code finds out round-trip time of an IP address by sending ping requests. You can provide the IP address and the number of packets to be sent as argument. Apart from that, it outputs the minimum, maximum, average round-trip time and percentage of packet loss to both console and a json file.